Alt Login Screen:  - Edited background
Anti Bot:
 - Updated funcraft anti bot

Auto Pot:
 - Can handle soups

Commands:
 - Bind added
 - HClip added
 - Name added

All 3D render modules:
 - Fixed colors and width

Notifications:
 - Added notification system

Click GUI:
 - Added Color Picker

Speed:
 - Added HypixelNew Speed (well, not rollbacking but i think its silent flagging)

Eagle:
 - Added

Via Version:
 - Updated to the last version

Scaffold:
 - Added Item Spoof modes (Pick, Keepspoof, Spoof)
 - Legit rotations updated (Now bypassing hypixel)
 - Keep Rotations added
 - Safe walk added

Main Menu:
 - Added clean main menu (screen below)

Phase:
 - Added clip mode

Target Strafe:
 - Way better
 - Predict positions (so there is lower chance to fall in the void)

+++ src/minecraft/net/minecraft/client/gui/Gui.java
@@ -185,7 +185,7 @@ public class Gui
     /**
      * Draws a textured rectangle at z = 0. Args: x, y, u, v, width, height, textureWidth, textureHeight
      */
-    public static void drawModalRectWithCustomSizedTexture(int x, int y, float u, float v, int width, int height, float textureWidth, float textureHeight)
+    public static void drawModalRectWithCustomSizedTexture(double x, double y, float u, float v, int width, int height, float textureWidth, float textureHeight)
     {
         float f = 1.0F / textureWidth;
         float f1 = 1.0F / textureHeight;
